+++ core/designsystem/src/main/java/com/iteneum/designsystem/components/phonenumbertext/PhoneNumberText.kt
@@ -54,55 +54,3 @@ fun mobileNumberFilter(text: AnnotatedString): TransformedText {
     }
     return TransformedText(annotatedString, phoneNumberOffsetTranslator)
 }
-
-/**
- * This function creates a phone number OutlinedTextField
- * @param modifier Set component modifier
- * @param showError This parameter determines whether the error is displayed or not
- * @param onPhoneChange Returns value typed
- *
- * @author Yaritza Moreno
- */
-@ExperimentalMaterial3Api
-@Composable
-fun PhoneNumberText(
-    modifier: Modifier = Modifier,
-    value: String,
-    imeAction: ImeAction = ImeAction.Next,
-    keyboardType: KeyboardType = KeyboardType.Phone,
-    isEnabled: Boolean = true,
-    showError: Boolean = false,
-    onPhoneChange: (String) -> Unit
-) {
-
-    var text by remember { mutableStateOf("") }
-    val maxChar = 10
-    val focusManager = LocalFocusManager.current
-
-    OutlinedTextField(
-        modifier = modifier.fillMaxWidth(),
-        label = { Text(text = "Contact phone") },
-        value = value,
-        onValueChange = {
-            text = it.take(maxChar)
-            if (it.length > maxChar) {
-                focusManager.moveFocus(FocusDirection.Down) // Or receive a lambda function
-            }
-            onPhoneChange(it)
-        },
-        textStyle = TextStyle(fontSize = 18.sp),
-        keyboardOptions = KeyboardOptions(
-            imeAction = imeAction,
-            keyboardType = keyboardType
-        ),
-        visualTransformation = { mobileNumberFilter(it) },
-        enabled = isEnabled,
-        colors = TextFieldDefaults.outlinedTextFieldColors(
-            textColor = Black,
-            focusedBorderColor = MaterialTheme.colorScheme.onPrimary,
-            focusedLabelColor = MaterialTheme.colorScheme.onPrimary,
-            cursorColor = MaterialTheme.colorScheme.onPrimary
-        )
-    )
-}
-
